updated quiz.py -> needs to be tested

+++ main.py
@@ -76,7 +76,7 @@ async def send_message(interaction: discord.Interaction, message: str):
 @bot.tree.command(name="set_points", description="changes the points of a user")
 @app_commands.describe(user="the user")
 @app_commands.describe(new_points="the new points")
-async def set_points(interaction: discord.Interaction, user: discord.User, new_points: float):
+async def set_points(interaction: discord.Interaction, user: discord.User, new_points: int):
     if interaction.user == quiz_master:
         await interaction.response.send_message(f"set points of {user.name} to {new_points}", ephemeral=True)
         await quiz.set_points(user, new_points)
